Created Wwise Project
Wwise Project is now set up, with a basic Survivor Attenuator (makes sounds quieter/more muted based on distance). Something to note: We will have to get in contact with people at Audiokinetic if we want a free license. Right now, we are limited to 200 sounds max until we do, which may not be an issue for a while.